+++ lib/src/controls/inputs/buttons/filled_button.dart
@@ -47,13 +47,13 @@ class FilledButton extends Button {
       if (theme.brightness.isDark) {
         return theme.accentColor.darker;
       } else {
-        theme.accentColor.lighter;
+        return theme.accentColor.lighter;
       }
     } else if (states.isHovering) {
       if (theme.brightness.isDark) {
         return theme.accentColor.dark;
       } else {
-        theme.accentColor.light;
+        return theme.accentColor.light;
       }
     }
     return theme.accentColor;