+++ lib/src/styles/theme.dart
@@ -176,6 +176,8 @@ class ThemeData with Diagnosticable {
   final Color scaffoldBackgroundColor;
   final Color acrylicBackgroundColor;
   final Color micaBackgroundColor;
+  final Color menuColor;
+  final Color cardColor;
 
   final Duration fasterAnimationDuration;
   final Duration fastAnimationDuration;
@@ -249,6 +251,8 @@ class ThemeData with Diagnosticable {
     required this.snackbarTheme,
     required this.pillButtonBarTheme,
     required this.bottomSheetTheme,
+    required this.menuColor,
+    required this.cardColor,
   });
 
   static ThemeData light() {
@@ -276,6 +280,8 @@ class ThemeData with Diagnosticable {
     Color? uncheckedColor,
     Color? checkedColor,
     Color? borderInputColor,
+    Color? menuColor,
+    Color? cardColor,
     Duration? fasterAnimationDuration,
     Duration? fastAnimationDuration,
     Duration? mediumAnimationDuration,
@@ -334,6 +340,8 @@ class ThemeData with Diagnosticable {
     borderInputColor ??= isLight
         ? const Color.fromRGBO(0, 0, 0, 0.4458)
         : const Color.fromRGBO(255, 255, 255, 0.5442);
+    menuColor ??= isLight ? const Color(0xFFf9f9f9) : const Color(0xFF2c2c2c);
+    cardColor ??= isLight ? const Color(0xFFf3f3f3) : const Color(0xFF2e2e2e);
     typography = Typography.standard(brightness: brightness)
         .merge(typography)
         .apply(fontFamily: fontFamily);
@@ -415,6 +423,8 @@ class ThemeData with Diagnosticable {
       snackbarTheme: snackbarTheme,
       pillButtonBarTheme: pillButtonBarTheme,
       bottomSheetTheme: bottomSheetTheme,
+      menuColor: menuColor,
+      cardColor: cardColor,
     );
   }
 
@@ -439,6 +449,7 @@ class ThemeData with Diagnosticable {
       uncheckedColor: Color.lerp(a.uncheckedColor, b.uncheckedColor, t)!,
       checkedColor: Color.lerp(a.checkedColor, b.checkedColor, t)!,
       borderInputColor: Color.lerp(a.borderInputColor, b.borderInputColor, t)!,
+      cardColor: Color.lerp(a.cardColor, b.cardColor, t)!,
       fasterAnimationDuration:
           lerpDuration(a.fasterAnimationDuration, b.fasterAnimationDuration, t),
       fastAnimationDuration:
@@ -479,6 +490,7 @@ class ThemeData with Diagnosticable {
           a.pillButtonBarTheme, b.pillButtonBarTheme, t),
       bottomSheetTheme:
           BottomSheetThemeData.lerp(a.bottomSheetTheme, b.bottomSheetTheme, t),
+      menuColor: Color.lerp(a.menuColor, b.menuColor, t)!,
     );
   }
 
@@ -498,6 +510,8 @@ class ThemeData with Diagnosticable {
     Color? uncheckedColor,
     Color? checkedColor,
     Color? borderInputColor,
+    Color? menuColor,
+    Color? cardColor,
     Duration? fasterAnimationDuration,
     Duration? fastAnimationDuration,
     Duration? mediumAnimationDuration,
@@ -543,6 +557,8 @@ class ThemeData with Diagnosticable {
       acrylicBackgroundColor:
           acrylicBackgroundColor ?? this.acrylicBackgroundColor,
       micaBackgroundColor: micaBackgroundColor ?? this.micaBackgroundColor,
+      menuColor: menuColor ?? this.menuColor,
+      cardColor: cardColor ?? this.cardColor,
       fasterAnimationDuration:
           fasterAnimationDuration ?? this.fasterAnimationDuration,
       fastAnimationDuration:
@@ -588,7 +604,9 @@ class ThemeData with Diagnosticable {
       ..add(ColorProperty('shadowColor', shadowColor))
       ..add(ColorProperty('scaffoldBackgroundColor', scaffoldBackgroundColor))
       ..add(ColorProperty('acrylicBackgroundColor', acrylicBackgroundColor))
-      ..add(ColorProperty('micaBackgroundColor', micaBackgroundColor));
+      ..add(ColorProperty('micaBackgroundColor', micaBackgroundColor))
+      ..add(ColorProperty('menuColor', menuColor))
+      ..add(ColorProperty('cardColor', cardColor));
     properties.add(EnumProperty('brightness', brightness));
     properties.add(DiagnosticsProperty<Duration>(
       'slowAnimationDuration',